add class UserService
