+++ client/src/js/index.js
@@ -1,6 +1,7 @@
 import "../css/styles.css";
 //модули
 import { UserController } from "./controllers/UserController";
+import { UserService } from "./services/UserService";
 import { UserModel } from "./models/UserModel";
 import { UserView } from "./views/UserView";
 
@@ -10,21 +11,33 @@ const fields = form.querySelectorAll("[name]");
 
 //экземпляры модулей
 const userModel = new UserModel();
+const userService = new UserService();
 const userView = new UserView(form.parentNode);
-const userController = new UserController(userView, userModel);
+const userController = new UserController(userView, userService, userModel);
+
+//обработчик при открытии страницы
+document.addEventListener(
+    "DOMContentLoaded",
+    () => {
+        if (sessionStorage.getItem("token")) {
+            sessionStorage.removeItem("token");
+        }
+    },
+    false
+);
 
 //обработчик отправки формы
 form.addEventListener("submit", submitForm);
 
 function submitForm(evt) {
-  evt.preventDefault();
+    evt.preventDefault();
 
-  userController.submit(evt.currentTarget);
+    userController.submit(evt.currentTarget);
 }
 
 //удаление класса ошибки из полей формы
 fields.forEach((field) => field.addEventListener("focus", clearError));
 
 function clearError(evt) {
-  userController.clearError(evt.currentTarget);
+    userController.clearError(evt.currentTarget);
 }
