+++ client/src/js/views/UserView.js
@@ -3,14 +3,17 @@ export class UserView {
         this.container = container;
     }
 
+    //вывод информации на страницу
     renderSucess(userName = "") {
         this.container.innerHTML = `${userName},<br>Вы успешно авторизованы.`;
     }
 
+    //вывод ошибки на страницу при неудачном запросе
     renderFailure(err) {
         this.container.innerHTML = `Произошла ошибка. <br> ${err}`;
     }
 
+    //отображение ошибки при валидации формы
     renderError(msg, attribute) {
         const element = this.container.querySelector(`[name="${attribute}"]`);
         const spanError = this.container.querySelector(`#error-${attribute}`);
@@ -25,6 +28,7 @@ export class UserView {
         element.blur();
     }
 
+    //очистка ошибки при валидации формы
     clearError(domObject) {
         const spanError = this.container.querySelector(`#error-${domObject.getAttribute("name")}`);
